tr2/objects/flare_item: move effect functions to object scope
This moves Flare_GenerateEffects and Flare_GenerateLight to the flare
item module along with related consts.
